% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{getAnnoyVersion}
\alias{getAnnoyVersion}
\title{Get the Annoy library version}
\usage{
getAnnoyVersion(compact = FALSE)
}
\arguments{
\item{compact}{Logical scalar indicating whether a compact
\code{\link{package_version}} should be returned.}
}
\value{
An integer vector containing the major, minor and patch version numbers;
or if \code{compact=TRUE}, a \code{\link{package_version}} object.
}
\description{
Get the version of the Annoy C++ library that RcppAnnoy was compiled with.
}
\author{
Aaron Lun
}
