% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoy.R
\docType{package}
\name{RcppAnnoy-package}
\alias{RcppAnnoy-package}
\alias{RcppAnnoy}
\alias{Rcpp_Annoy}
\title{Rcpp bindings for the Annoy C++ library for approximate nearest neighbors.}
\description{
Annoy is a small library written to provide fast and memory-efficient
nearest neigbor lookup from a possibly static index which can be
shared across processes.
}
\details{
Details about Annoy are available at the reference listed below.
}
\references{
\url{https://github.com/spotify/annoy}
}
\author{
Dirk Eddelbuettel for the R interface; Erik Bernhardsson for Annoy itself.

Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\keyword{package}
