\name{RcppAnnoy-package}
\alias{RcppAnnoy-package}
\alias{RcppAnnoy}
\alias{Rcpp_Annoy}
\docType{package}
\title{
  Rcpp bindings for the Annoy C++ library for approximate nearest neighbors.
}
\description{
  Annoy is a small library written to provide fast and memory-efficient
  nearest neigbor lookup from a possibly static index which can be
  shared across processes.
}
\details{
  Details about Annoy are available at the reference listed below.
}
\author{
  Dirk Eddelbuettel for the R interface; Erik Bernhardsson for Annoy itself.

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\references{
  \url{https://github.com/spotify/annoy}
}
\keyword{package}
\examples{
  # Optional simple examples of the most important functions
}
